Fix the combined graph, to avoid some 'combsched' errors
The execution order and urgency graphs are made of edges between rules
(and methods) as a whole.  In the combined graph, there are separate
nodes for the Sched and Exec of each rule (and method).  When adding
execution order edges to the combined graph, we had been creating
Exec-to-Exec edges; but an Exec-to-Sched edge is needed, when the
conflicting uses are in the rule's predicate.  We now create the
edges in this form, by keeping tracking (in the conflict map) of
whether any method uses were in the predicate.  Specifically, the
CUse edge type has a new Bool field, which is True if any are pred
uses.  This fixes bug 1706 in Bluespec Inc's bug database.