Rough attempt at fixing issue 633
This does kind inference in 'getCls' to try to properly convert the
untyped CSyntax for superclasses into kind/typechecked CSyntax.
But use of the typeclass still doesn't work.  Need to confirm that
the output of 'getCls' is correct; if so, then check where the
superclass info is being used, and see if something needs to be
done there (such as replacing the dependent type variables in the
superclasses with fresh variables).  Also, 'getCls' will need to do
error checking that the superclasses don't have any non-dependent
variables (and report an ambiguous explicit context error).